# api/openapi/controlplane.v1.yaml
openapi: 3.1.0
info:
  title: AgentGate Control Plane API
  version: "1.0.0"
  summary: Agent registration, workload identity issuance, and the promotion gate.
  description: |
    The control plane is the trust plane: it owns the agent inventory, issues and rotates workload
    identity, and runs the promotion gate that decides whether a version may serve traffic in an
    environment.

    It is deliberately **off the request path**. The gateway verifies tokens against a cached key
    set and enforces on signed claims, so a control-plane outage stops new identities from being
    issued but does not stop agents that are already running. That separation is the single most
    important availability property of the platform, and it is the reason this API is documented
    separately from `gateway.v1.yaml` rather than folded into it.

    ## Two surfaces, two authentication models

    * **The OAuth2 surface** — `/.well-known/*` and `/oauth2/token` — is standards-shaped so that
      an enterprise identity provider or an API management layer can validate AgentGate tokens
      without a bespoke integration. `/oauth2/token` is `application/x-www-form-urlencoded` and
      returns OAuth2-style `{"error", "error_description"}` bodies, not problem+json.
    * **The management surface** — `/api/v1/*` — is JSON and returns RFC 9457 problem+json. It is
      authenticated with a bearer token and authorised by scope: `agents:register`,
...

...
      promotion day.

    ## Compatibility

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /oauth2/token:
    post:
      tags: [token]
      operationId: issueToken
      summary: Issue an AgentGate access token
      description: |
        Two grants are supported. Both are `application/x-www-form-urlencoded`.

        ### `urn:ietf:params:oauth:grant-type:token-exchange` (preferred)

        RFC 8693. The agent's runtime has already proved who it is — a Kubernetes projected service
        account token, an Azure managed identity token, an IAM Roles Anywhere or SPIFFE JWT-SVID —
        and the control plane converts that proof into an AgentGate identity. **No shared secret
        ever exists.**

        The subject token's issuer must be one of the configured trusted issuers, its signature
        must verify, and its subject must correspond to the agent being requested. Correspondence
        is checked against the agent's `workload_subject` label when one is recorded, and otherwise
        by requiring the agent name to appear in the workload subject. Without this check any
        workload in the cluster could obtain any agent's identity.

        Tokens from this grant carry `attestation=workload-identity`, which is what the production
        promotion gate requires.

        ### `client_credentials` (fallback)

        For runtimes that cannot do federated workload identity. Credentials may be sent as
        `client_id`/`client_secret` form fields or as HTTP Basic authentication. An unknown client
        and a wrong secret produce the same `invalid_client` response: distinguishing them is an
        enumeration oracle.

        Tokens from this grant carry `attestation=client-secret`, which the production promotion
        gate refuses.

        ### Version selection

        `agent_version` selects a specific version. When omitted, the version currently `active` in
        the requested `env` is used. If no version is active there, the request is refused with
        `403 invalid_grant` — a version that has not been promoted cannot obtain an identity for
        that environment, which is what makes the promotion gate load-bearing rather than advisory.

        Responses carry `Cache-Control: no-store`: a token is a credential and must never be held
        by an intermediary.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequestForm'
            examples:
              tokenExchange:
                summary: Token exchange from a Kubernetes projected service account token
                value:
                  grant_type: urn:ietf:params:oauth:grant-type:token-exchange
                  subject_token: eyJhbGciOiJSUzI1NiIsImtpZCI6ImsxIn0.eyJpc3MiOiJodHRwczovL2t1YmVybmV0ZXMuZGVmYXVsdC5zdmMifQ.sig
                  subject_token_type: urn:ietf:params:oauth:token-type:jwt
                  agent_identity: agent://fsclient/payments-risk/dispute-triage
                  env: prod
                  agent_version: 2.4.1
              clientCredentials:
                summary: Client credentials
                value:
                  grant_type: client_credentials
                  client_id: cli_01j8z9x2qk4m7p3r5t9v
                  client_secret: ags_Q2xpZW50U2VjcmV0RXhhbXBsZVZhbHVl
                  env: staging
      responses:
        '200':
          description: An access token.
          headers:
            Cache-Control:
              description: Always `no-store`.
              schema:
                type: string
            Pragma:
              description: Always `no-cache`.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
              examples:
                issued:
                  value:
                    access_token: eyJhbGciOiJSUzI1NiIsImtpZCI6ImstMjAyNi0wOCJ9.eyJzdWIiOiJhZ2VudDovLyJ9.sig
                    token_type: Bearer
                    expires_in: 900
                    scope: models:invoke models:embed
                    issued_at: 1774526521
                    agent_id: agt_01j8z9x2qk4m7p3r5t9v
                    env: prod
                    agent_version: 2.4.1
        '400':
          description: |
            `invalid_request` — a required form field is missing or the body could not be parsed.
            `invalid_grant` — the subject token could not be verified, or the agent is not
            registered. `unsupported_grant_type` — `grant_type` is neither supported grant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
              examples:
                missingFields:
                  value:
                    error: invalid_request
                    error_description: subject_token and agent_identity are required
                unsupportedGrant:
                  value:
                    error: unsupported_grant_type
                    error_description: supported grants are client_credentials and urn:ietf:params:oauth:grant-type:token-exchange
        '401':
          description: '`invalid_client` — client authentication failed.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
              examples:
                badClient:
                  value:
                    error: invalid_client
                    error_description: client authentication failed
        '403':
          description: |
            `invalid_grant` — the subject token does not correspond to the requested agent, or no
            version of the agent is promoted for the requested environment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
              examples:
                notPromoted:
                  value:
                    error: invalid_grant
                    error_description: no active version of this agent is promoted for prod
                subjectMismatch:
                  value:
                    error: invalid_grant
                    error_description: the subject token does not correspond to this agent
        '500':
          description: '`server_error` — token minting failed for a reason that is not the caller''s.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
  /api/v1/agents:
    post:
      tags: [agents]
      operationId: registerAgent
      summary: Register an agent version
      description: |
        Called by CI on every deployment. **Idempotent on (identity, env, version)**: re-running a
        pipeline does not create a second agent or a second credential.

        A version registered into `dev` becomes `active` immediately — an engineer must be able to
        run their agent against the gateway the same day they write it, or they will build
        something that bypasses it. A version registered into `staging` or `prod` becomes
        `registered` and must be promoted.

        Ownership is mutable through this endpoint **except** the owning team: changing
        `owner.team` is refused with `409 conflict`, because chargeback history would otherwise
        move silently between cost centres. That transfer is a platform action.

        `warnings` reports problems that do not block registration but will block promotion, so a
        team learns about them one deployment early rather than on the day they try to ship.

        Requires the `agents:register` scope. Unknown JSON fields are rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
            examples:
              federated:
                summary: Registration with federated workload identity (no secret)
                value:
                  identity: agent://fsclient/payments-risk/dispute-triage
                  display_name: Dispute Triage Agent
                  description: Classifies inbound card disputes and routes them to the right queue.
                  owner:
                    team: payments-risk
                    email: payments-risk@client.example
                    oncall: PD-PAYRISK
                    cost_center: CC-4471
                    manager: a.okafor@client.example
                  runtime: aks
                  framework: langgraph
                  data_classification: confidential
                  requested_pools: [general-chat, long-context]
                  quota:
                    tokens_per_minute: 120000
                    requests_per_minute: 600
                    monthly_token_budget: 900000000
                    max_concurrent: 40
                  version: 2.4.1
                  env: staging
                  image: registry.client.example/payments-risk/dispute-triage:2.4.1
                  commit_sha: 9f21c4ade1b8f0c3d5a7e2419b6c8d0f3a1e5b7c
                  security_review_ref: CHG0048221
                  labels:
                    workload_subject: system:serviceaccount:payments-risk:dispute-triage
              withCredential:
                summary: Registration asking for a client secret (discouraged)
                value:
                  identity: agent://fsclient/payments-risk/dispute-triage
                  display_name: Dispute Triage Agent
                  owner:
                    team: payments-risk
                    email: payments-risk@client.example
                    oncall: PD-PAYRISK
                    cost_center: CC-4471
                  runtime: vm
                  data_classification: internal
                  requested_pools: [general-chat]
                  quota:
                    tokens_per_minute: 20000
                    requests_per_minute: 60
                    monthly_token_budget: 50000000
                  version: 0.9.0
                  env: dev
                  issue_credential: true
      responses:
        '200':
          description: |
            The agent record after registration. `client_secret` is present only when
            `issue_credential` was true; it is shown **once** and never stored in a form that can
            be read back. The response then carries `Cache-Control: no-store`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResult'
              examples:
                registered:
                  value:
                    agent:
                      agent_id: agt_01j8z9x2qk4m7p3r5t
                      identity: agent://fsclient/payments-risk/dispute-triage
                      display_name: Dispute Triage Agent
                      owner:
                        team: payments-risk
                        email: payments-risk@client.example
                        oncall: PD-PAYRISK
                        cost_center: CC-4471
                      runtime: aks
                      framework: langgraph
                      data_classification: confidential
                      requested_pools: [general-chat, long-context]
                      granted_pools: [general-chat]
                      quota:
                        tokens_per_minute: 120000
                        requests_per_minute: 600
                        monthly_token_budget: 900000000
                      versions:
                        - version: 2.4.1
                          env: staging
                          state: registered
                          image: registry.client.example/payments-risk/dispute-triage:2.4.1
                          commit_sha: 9f21c4ade1b8f0c3d5a7e2419b6c8d0f3a1e5b7c
                          registered_at: "2026-08-26T09:14:02Z"
                          registered_by: ci@client.example
                      security_review_ref: CHG0048221
                      created_at: "2026-05-02T11:00:00Z"
                      updated_at: "2026-08-26T09:14:02Z"
                    version: 2.4.1
                    env: staging
                    state: registered
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}

    get:
      tags: [agents]
      operationId: listAgents
      summary: List registered agents
      description: |
        Requires the `fleet:read` scope. `tenant` and `team` are applied in the query; `env` and
        `q` are applied in the application, so the same semantics hold for every store backend and
        so a filter change does not need a database migration.
      parameters:
        - name: tenant
          in: query
          schema: {type: string}
          example: fsclient
        - name: team
          in: query
          schema: {type: string}
          example: payments-risk
        - name: env
          in: query
          description: Return agents that have at least one version in this environment.
          schema:
            type: string
            enum: [dev, staging, prod]
        - name: q
          in: query
          description: Case-insensitive substring match over identity and display name.
          schema: {type: string}
      responses:
        '200':
          description: Matching agents, ordered by identity.
          content:
            application/json:
              schema:
                type: object
                required: [agents, count]
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  count:
                    type: integer
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/agents/{id}:
    get:
      tags: [agents]
      operationId: getAgent
      summary: Show one agent
      description: Requires the `fleet:read` scope. `id` is the `agent_id`, not the identity URI.
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: The agent record, including every version and its last gate snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/agents/{id}/credentials:
    post:
      tags: [credentials]
      operationId: issueCredential
      summary: Issue a client secret
      description: |
        Mints a client credential for an agent. Only the SHA-256 hash of the secret is stored in
        the registry; the secret itself goes to the enterprise vault, and the plaintext is returned
        **once**.

        Prefer token exchange with federated workload identity. Issuing a secret is logged at WARN,
        and the resulting attestation is one the production promotion gate refuses.

        Requires the `agents:register` scope. The request has no body.
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '201':
          description: The credential. The secret is shown once.
          headers:
            Cache-Control:
              description: Always `no-store`.
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedCredential'
              examples:
                issued:
                  value:
                    client_id: cli_01j8z9x2qk4m7p3r5t9v
                    client_secret: ags_Q2xpZW50U2VjcmV0RXhhbXBsZVZhbHVl
                    expires_at: "2026-11-24T09:14:02Z"
                    warning: this secret is shown once; prefer token exchange with federated workload identity
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/agents/{id}/credentials/{clientID}/rotate:
    post:
      tags: [credentials]
      operationId: rotateCredential
      summary: Rotate a client secret with an overlap window
      description: |
        Issues a new secret while leaving the previous one valid for the overlap window, so a
        running agent picks up the new secret on its next restart rather than at the instant of
        rotation. A rotation with no overlap is an outage scheduled for whenever the agent next
        needs a token.

        Requires the `agents:register` scope. The request has no body. Refused with `409 conflict`
        when the credential does not belong to the named agent.
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - name: clientID
          in: path
          required: true
          description: The client id being rotated.
          schema: {type: string}
          example: cli_01j8z9x2qk4m7p3r5t9v
        - name: overlap
          in: query
          required: false
          description: |
            How long the previous secret stays valid, as a Go duration (`24h`, `72h`, `30m`).
            Defaults to `24h`. An unparseable value falls back to the default rather than failing.
          schema:
            type: string
            default: 24h
          example: 72h
      responses:
        '201':
          description: The new credential, and when the previous one stops working.
          headers:
            Cache-Control:
              description: Always `no-store`.
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotatedCredential'
              examples:
                rotated:
                  value:
                    client_id: cli_02k9a0y3rl5n8q4s6u
                    client_secret: ags_Um90YXRlZFNlY3JldEV4YW1wbGVWYWx1ZQ
                    previous_valid_until: "2026-08-29T09:14:02Z"
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/v1/promotions/{id}:
    get:
      tags: [promotions]
      operationId: getPromotion
      summary: Show one promotion request
      description: |
        The auditable record: who asked, when, what the gate observed, who approved, and when it
        was applied. Requires the `fleet:read` scope.
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string}
          example: prm_1f0c8b3a-7d21-4e55-9a6b-0c2e8f4d1a93
      responses:
        '200':
          description: The promotion request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionRequest'
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A bearer token identifying a human or a CI principal, verified by the control plane's admin
        verifier and authorised by scope. Registration is called by CI, promotion by an engineer or
        a pipeline, and approval by a human; all three authenticate the same way and are
        distinguished by scope, so the audit record always names a principal rather than "the
        system". The audit actor is the token's `owner_email` claim when present, otherwise its
        subject.
  parameters:
    AgentId:
      name: id
      in: path
      required: true
      description: The `agent_id`, for example `agt_01j8z9x2qk4m7p3r5t`. Not the identity URI.
      schema: {type: string}
      example: agt_01j8z9x2qk4m7p3r5t
...